Avoid removing the whole file system
If scripts are used by hand and
WORKSPACE is not set then rm -rf /*
will remove the filesystem.

+++ jjb/3rd_party_ci/download-netvirt-artifact.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -e
 
+if [ -z ${WORKSPACE} ]; then
+  echo "WORKSPACE is unset. Please do so."
+  exit 1
+fi
 # wipe the WORKSPACE
 /bin/rm -rf $WORKSPACE/*
